requirements.txt
farm-haystack==1.16.1
fastapi==0.95.1
fastapi-pagination==0.12.3
spacy==3.5.2
uvicorn==0.22.0